Remove `client_id_scheme` from Verifier Metadata
There is currently some ambiguity around `client_id_scheme`. This parameter appears in the request object AND the client metadata, which also appears in the request object. There is no text in the specification to define what a wallet should do if it receives a rquest object where these two values do not match.

As I see it there are 3 ways forwards:
1. Remove `client_id_scheme` from Verifier Metadata.
2. Remove `client_id_scheme` from the authz request object, AND make it so that either `client_metadata` or `client_metadata_uri` are always provided in the request object.
3. Add text to describe what a wallet should do if it receives a request object with mismatching `client_id_scheme`.

I have raised this PR with option 1.