Update the text about the response returned by the browser API
This makes it consistent with the latest position from the WICG group, in
particular as shown here:

https://github.com/WICG/digital-credentials/pull/141/files

i.e. the VP response is inside the 'data' member within the response
object.

Also fix where presentation_submission was shown as a string in one example
and add an example for an encrypted response for completeness.

closes #201