Fix GitHub action that publishes compiled version spec from main
The upgrade of the download-artifact action from the version that github
disabled broke other things. Upgrade various gitlab actions to latest
versions to resolve this and other (non-fatal) warnings we were getting
in the frontend, e.g. things like:

The following artifacts were uploaded using a version of actions/upload-artifact that is scheduled for deprecation: "github-pages".