Fix compile error with vp_query ref (with Kristina's permission)
I'd made a suggestion that added a reference to the new query language,
but that stops the spec compiling because the new query language isn't
merged yet and won't be for a few days. To let the spec compile, just
remove that reference. We can always add it back later.